Use PyTorch's batch normalization.
The batch normalization for decoder is only done using training, it won't be done during the validation and test phase so no need to worry about momentum and running means and running variances. Reuse the means and variance calculated on z_pre to batch normalize the input for decoder. Note that this will only happen during the training phase. The decoder is turned off during validation and testing phase.